monkey patch for my current environment: cygwin(x86_64, 3.04) + libhdf5(1.10.2)
ruby: 2.3.6p384
narray: 0.6.1.2

+++ lib/hdf5.rb
@@ -49,7 +49,7 @@ module Hdf5
         :int
       end
       def hid_t
-        :int
+        :int64
       end
       def hbool_t
         :uint
